fix: added abstract class on top of datastoreUtils and refactored code for this
